+++ packages/compiler/src/index.ts
@@ -1,4 +1,5 @@
 import compile from './compile';
+export { default as getExports } from './getExports';
 
 export * from './util';
 export type { TransformOptions } from './transform';