+++ packages/runtime/test/execute/compile-plan.test.ts
@@ -433,6 +433,43 @@ test('throw for a syntax error on a job edge', (t) => {
   }
 });
 
+test('globals should be undefined if no globals are passed', (t) => {
+  const plan: ExecutionPlan = {
+    workflow: {
+      steps: [{ expression: 'x' }, { expression: 'y' }],
+    },
+    options: {},
+  };
+  const compiled = compilePlan(plan);
+  t.is(compiled.workflow.globals, undefined);
+});
+
+test('globals should be undefined if non-string is passed', (t) => {
+  const plan: ExecutionPlan = {
+    workflow: {
+      globals: { some: 'val' } as any,
+      steps: [{ expression: 'x' }, { expression: 'y' }],
+    },
+    options: {},
+  };
+  const compiled = compilePlan(plan);
+  t.is(compiled.workflow.globals, undefined);
+});
+
+test('global functions should perform no transformation inline code', (t) => {
+  const GLOBAL_FUNCTIONS =
+    'export const cleaner = (name) => name.replace("a", "b");';
+  const plan: ExecutionPlan = {
+    workflow: {
+      globals: GLOBAL_FUNCTIONS,
+      steps: [{ expression: 'x' }, { expression: 'y' }],
+    },
+    options: {},
+  };
+  const compiled = compilePlan(plan);
+  t.is(compiled.workflow.globals, GLOBAL_FUNCTIONS);
+});
+
 test('throw for multiple errors', (t) => {
   const plan = {
     workflow: {