+++ integration-tests/worker/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @openfn/integration-tests-worker
 
+## 1.0.89
+
+### Patch Changes
+
+- Updated dependencies [9d4ece3]
+  - @openfn/ws-worker@1.14.0
+  - @openfn/engine-multi@1.6.7
+  - @openfn/lightning-mock@2.2.1
+
 ## 1.0.88
 
 ### Patch Changes