+++ packages/runtime/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/runtime
 
+## 1.7.0
+
+### Minor Changes
+
+- 9d4ece3: Add support for global functions in execution plan
+
 ## 1.6.4
 
 ### Patch Changes